requirements.txt
streamlit
langchain-openai
langchain-groq
langchain-google-genai
langchain-core
python-dotenv
scikit-learn
numpy
pypdf
duckduckgo-search
requests
